+++ database/migrations/2018_12_07_075115_relasi_role_user.php
@@ -14,7 +14,7 @@ class RelasiRoleUser extends Migration
     public function up()
     {
         Schema::table('users', function($table) {
-            $table->unsignedInteger('role_id')->default(2);
+            $table->unsignedTinyInteger('role_id')->default(2);
             $table->foreign('role_id')->references('id')->on('roles');
         });
     }
