+++ database/migrations/2019_01_07_071307_set_relasi_customer_dan_cart.php
@@ -14,7 +14,7 @@ class SetRelasiCustomerDanCart extends Migration
     public function up()
     {
         Schema::table('carts', function($table) {
-            $table->unsignedInteger('customer_id');
+            $table->unsignedTinyInteger('customer_id');
             $table->foreign('customer_id')->references('id')->on('customers');
         });
     }
