+++ database/migrations/2019_01_30_131009_set_relasi_cart_customer_dan_user.php
@@ -14,17 +14,17 @@ class SetRelasiCartCustomerDanUser extends Migration
     public function up()
     {
         Schema::table('payments', function($table) {
-            $table->unsignedInteger('cart_id');
+            $table->unsignedTinyInteger('cart_id');
             $table->foreign('cart_id')->references('id')->on('carts');
         });
 
         Schema::table('payments', function($table) {
-            $table->unsignedInteger('customer_id');
+            $table->unsignedTinyInteger('customer_id');
             $table->foreign('customer_id')->references('id')->on('customers');
         });
 
         Schema::table('payments', function($table) {
-            $table->unsignedInteger('user_id');
+            $table->unsignedSmallInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
         });
     }
